requirements.txt
python-telegram-bot
python-dotenv
openai
watchdog
